+++ components-core/src/main/java/com/adyen/checkout/components/core/internal/ui/model/SessionParams.kt
@@ -12,10 +12,19 @@ import androidx.annotation.RestrictTo
 import com.adyen.checkout.components.core.Amount
 import java.util.Locale
 
+/**
+ * Object that holds values set during sessions setup call.
+ * [SessionParams] values by default should have higher priority than values set in client side configurations.
+ * Otherwise it can cause server error, since specific configuration is not enabled, but it is being used.
+ *
+ * Only for some specific cases, if they do not cause server error, client side configurations can have higher priority
+ * than [SessionParams] values.
+ */
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 data class SessionParams(
     val enableStoreDetails: Boolean?,
     val installmentConfiguration: SessionInstallmentConfiguration?,
+    val showRemovePaymentMethodButton: Boolean?,
     val amount: Amount?,
     val returnUrl: String?,
     val shopperLocale: Locale?,