+++ drop-in/src/main/java/com/adyen/checkout/dropin/SessionDropInService.kt
@@ -12,6 +12,7 @@ import com.adyen.checkout.components.core.ActionComponentData
 import com.adyen.checkout.components.core.OrderRequest
 import com.adyen.checkout.components.core.OrderResponse
 import com.adyen.checkout.components.core.PaymentComponentState
+import com.adyen.checkout.components.core.StoredPaymentMethod
 import com.adyen.checkout.core.AdyenLogLevel
 import com.adyen.checkout.core.Environment
 import com.adyen.checkout.core.internal.data.api.HttpClientFactory
@@ -88,7 +89,7 @@ open class SessionDropInService : BaseDropInService(), SessionDropInServiceInter
                 is SessionCallResult.Payments.Action -> DropInServiceResult.Action(result.action)
                 is SessionCallResult.Payments.Error ->
                     DropInServiceResult.Error(
-                        errorDialog = ErrorDialog(),
+                        errorDialog = ErrorDialog(message = getString(R.string.payment_failed)),
                         reason = result.throwable.message,
                         dismissDropIn = true,
                     )
@@ -97,7 +98,7 @@ open class SessionDropInService : BaseDropInService(), SessionDropInServiceInter
                 is SessionCallResult.Payments.NotFullyPaidOrder -> updatePaymentMethods(result.result.order)
                 is SessionCallResult.Payments.RefusedPartialPayment ->
                     DropInServiceResult.Error(
-                        errorDialog = ErrorDialog(),
+                        errorDialog = ErrorDialog(message = getString(R.string.payment_failed)),
                         reason = "Payment was refused while making a partial payment",
                     )
 
@@ -123,7 +124,7 @@ open class SessionDropInService : BaseDropInService(), SessionDropInServiceInter
                 is SessionCallResult.Details.Action -> DropInServiceResult.Action(result.action)
                 is SessionCallResult.Details.Error ->
                     DropInServiceResult.Error(
-                        errorDialog = ErrorDialog(),
+                        errorDialog = ErrorDialog(message = getString(R.string.payment_failed)),
                         reason = result.throwable.message,
                         dismissDropIn = true,
                     )
@@ -150,7 +151,7 @@ open class SessionDropInService : BaseDropInService(), SessionDropInServiceInter
             val dropInServiceResult = when (result) {
                 is SessionCallResult.Balance.Error ->
                     BalanceDropInServiceResult.Error(
-                        errorDialog = ErrorDialog(),
+                        errorDialog = ErrorDialog(message = getString(R.string.payment_failed)),
                         reason = result.throwable.message,
                     )
 
@@ -175,7 +176,7 @@ open class SessionDropInService : BaseDropInService(), SessionDropInServiceInter
             val dropInServiceResult = when (result) {
                 is SessionCallResult.CreateOrder.Error ->
                     OrderDropInServiceResult.Error(
-                        errorDialog = ErrorDialog(),
+                        errorDialog = ErrorDialog(message = getString(R.string.payment_failed)),
                         reason = result.throwable.message,
                         dismissDropIn = true,
                     )
@@ -203,7 +204,7 @@ open class SessionDropInService : BaseDropInService(), SessionDropInServiceInter
             val dropInServiceResult = when (result) {
                 is SessionCallResult.CancelOrder.Error ->
                     DropInServiceResult.Error(
-                        errorDialog = ErrorDialog(),
+                        errorDialog = ErrorDialog(message = getString(R.string.unknown_error)),
                         reason = result.throwable.message,
                     )
 
@@ -222,6 +223,27 @@ open class SessionDropInService : BaseDropInService(), SessionDropInServiceInter
         }
     }
 
+    override fun onRemoveStoredPaymentMethod(storedPaymentMethod: StoredPaymentMethod) {
+        launch {
+            val storedPaymentMethodId = storedPaymentMethod.id.orEmpty()
+            val result = sessionInteractor.removeStoredPaymentMethod(storedPaymentMethodId)
+
+            val serviceResult = when (result) {
+                SessionCallResult.RemoveStoredPaymentMethod.Successful ->
+                    RecurringDropInServiceResult.PaymentMethodRemoved(
+                        storedPaymentMethodId,
+                    )
+
+                is SessionCallResult.RemoveStoredPaymentMethod.Error -> RecurringDropInServiceResult.Error(
+                    errorDialog = ErrorDialog(message = getString(R.string.unknown_error)),
+                    reason = result.throwable.message,
+                )
+            }
+
+            sendRecurringResult(serviceResult)
+        }
+    }
+
     private suspend fun updatePaymentMethods(order: OrderResponse? = null): DropInServiceResult {
         return when (val result = sessionInteractor.updatePaymentMethods(order)) {
             is SessionCallResult.UpdatePaymentMethods.Successful -> DropInServiceResult.Update(
@@ -231,7 +253,7 @@ open class SessionDropInService : BaseDropInService(), SessionDropInServiceInter
 
             is SessionCallResult.UpdatePaymentMethods.Error ->
                 DropInServiceResult.Error(
-                    errorDialog = ErrorDialog(),
+                    errorDialog = ErrorDialog(message = getString(R.string.payment_failed)),
                     reason = result.throwable.message,
                     dismissDropIn = true,
                 )